refactor: move game and make to core, full context
