+++ src/core/context.ts
@@ -8,8 +8,16 @@ type ParametersOmitFirst<T extends (x: any, ...args: any) => any> = T extends
 type ParametersOmitFirst2<T extends (x: any, y: any, ...args: any) => any> =
     T extends (x: any, y: any, ...args: infer P) => any ? P : never;
 
-type Wrapper<T> = {
+type ParametersOmitFirst4<T extends (x: any, y: any, ...args: any) => any> =
+    T extends (x: any, y: any, a: any, b: any, ...args: infer P) => any ? P
+        : never;
+
+export type Wrapper<T> = {
     current: T | null;
+};
+
+export type StrictWrapper<T> = {
+    current: T;
 };
 
 // This works for methods, component that uses the components, this means
@@ -70,6 +78,23 @@ export const withGfx = <
     };
 };
 
+type Game = any;
+
+export const withGame = <
+    T extends (game: Game, ...args: any[]) => any,
+    R extends ReturnType<T>,
+>(
+    func: T,
+) => {
+    return (game: Game): (
+        ...args: ParametersOmitFirst<T>
+    ) => R => {
+        return (...args: ParametersOmitFirst<T>) => {
+            return func(game, ...args);
+        };
+    };
+};
+
 export const withAppAndKaboom = <
     T extends (app: App, k: KaboomCtx, ...args: any[]) => any,
     R extends ReturnType<T>,
@@ -104,6 +129,39 @@ export const withGfxAndKaboom = <
             }
 
             return func(gfx, k.current, ...args);
+        };
+    };
+};
+
+// This context wraps some functions and wrappers even if they doesn't exist.
+export const withFullCtx = <
+    T extends (
+        app: App,
+        gfx: AppGfxCtx,
+        game: Wrapper<Game>,
+        k: StrictWrapper<KaboomCtx>,
+        ...args: any[]
+    ) => any,
+    R extends ReturnType<T>,
+>(
+    func: T,
+) => {
+    return (
+        app: App,
+        gfx: AppGfxCtx,
+        game: Wrapper<Game>,
+        k: Wrapper<KaboomCtx>,
+    ): (
+        ...args: ParametersOmitFirst4<T>
+    ) => R => {
+        return (...args: ParametersOmitFirst4<T>) => {
+            return func(
+                app,
+                gfx,
+                game,
+                k as StrictWrapper<KaboomCtx>,
+                ...args,
+            );
         };
     };
 };
